Added timer and player classes.
The timer class keeps track of the current time, makes sure it
increments at the right time and can also return the current time in
minutes + seconds.
The player class keeps track of some simple properties at this point
(such as name, score, highscore)

Also added that the console will log the current time + current score.
When a pole is hit, it will set the highscore and, when it’s better
than the previous highscore, it will also log it.